pyproject.toml
[project]
name = "fancryrag"
version = "0.1.0"
description = "Neo4j GraphRAG baseline configured for Neo4j vector storage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "neo4j-graphrag[experimental,openai]",
    "neo4j>=5.18,<6",
    "python-dotenv>=1.0.1,<2",
    "fastmcp>=0.3.0",
    "pydantic>=2.7,<3",
]

[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7",
    "pytest>=8.3",
    "pytest-asyncio>=0.23",
    "ruff>=0.6.4",
]

[tool.pytest.ini_options]
markers = [
    "integration: Tests that require Docker Compose and container orchestration",
]

[dependency-groups]
dev = [
    "pyyaml>=6.0.3",
]
